desktop-app/requirements.txt
PyQt5>=5.15.0
mss>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
pyautogui>=0.9.54
requests>=2.31.0
pynput>=1.7.6
screeninfo>=0.8.1
Flask>=2.3.0
flask-cors>=4.0.0

